Completely remove WaitForChannel
