+++ packages/wallet/src/redux/channel-state/__tests__/channel-state.test.ts
@@ -3,6 +3,9 @@ import * as actions from '../../actions';
 import * as scenarios from '../../__tests__/test-scenarios';
 import * as channelState from '../reducer';
 import * as states from '../state';
+import { itTransitionsToChannelStateType, itSendsThisMessage } from '../../__tests__/helpers';
+import * as SigningUtil from '../../../utils/signing-utils';
+import { validationFailure, SIGNATURE_FAILURE } from 'magmo-wallet-client';
 
 const {
   initializingChannelState: initializingChannels,
@@ -10,6 +13,8 @@ const {
   channelId,
   asAddress,
   asPrivateKey,
+  bsAddress,
+  bsPrivateKey,
   preFundCommitment1,
   preFundCommitment2,
 } = scenarios;
@@ -100,5 +105,72 @@ describe('when the channel is part of the channelState', () => {
       channelState.channelStateReducer(state, action);
       expect(mock).toBeCalledTimes(0);
     });
+  });
+});
+
+describe('when the channel is initializing', () => {
+  describe('when we send in a PreFundSetupA', () => {
+    // preFundSetupA is A's move, so in this case we need to be player A
+    const initializingChannelsA = {
+      [asAddress]: { address: asAddress, privateKey: asPrivateKey },
+    };
+    const state = { ...defaults, initializingChannels: initializingChannelsA };
+    const action = actions.channel.ownCommitmentReceived(preFundCommitment1);
+
+    const updatedState = channelState.channelStateReducer(state, action);
+    const initializedChannel = { state: states.getChannel(updatedState.state, channelId) };
+
+    itTransitionsToChannelStateType(states.WAIT_FOR_PRE_FUND_SETUP, initializedChannel);
+  });
+
+  describe('when an opponent sends a PreFundSetupA', () => {
+    // preFundSetupA is A's move, so in this case we need to be player B
+    const initializingChannelsB = {
+      [bsAddress]: { address: bsAddress, privateKey: bsPrivateKey },
+    };
+    const state = { ...defaults, initializingChannels: initializingChannelsB };
+    const action = actions.channel.opponentCommitmentReceived(preFundCommitment1, 'sig');
+    const validateMock = jest.fn().mockReturnValue(true);
+    Object.defineProperty(SigningUtil, 'validCommitmentSignature', { value: validateMock });
+
+    const updatedState = channelState.channelStateReducer(state, action);
+    const initializedChannel = { state: states.getChannel(updatedState.state, channelId) };
+
+    itTransitionsToChannelStateType(states.WAIT_FOR_PRE_FUND_SETUP, initializedChannel);
+  });
+
+  describe('when an opponent sends a PreFundSetupA but the signature is bad', () => {
+    const initializingChannelsA = {
+      [asAddress]: { address: asAddress, privateKey: asPrivateKey },
+    };
+    const state = { ...defaults, initializingChannels: initializingChannelsA };
+    const action = actions.channel.opponentCommitmentReceived(
+      preFundCommitment1,
+      'not-a-signature',
+    );
+    const validateMock = jest.fn().mockReturnValue(false);
+    Object.defineProperty(SigningUtil, 'validCommitmentSignature', { value: validateMock });
+
+    const updatedState = channelState.channelStateReducer(state, action);
+    const initializedChannel = states.getChannel(updatedState.state, channelId);
+
+    expect(initializedChannel).toBeUndefined(); // it doesn't transition
+
+    itSendsThisMessage(updatedState, validationFailure('InvalidSignature'));
+  });
+
+  describe('when we send in a a non-PreFundSetupA', () => {
+    const initializingChannelsA = {
+      [asAddress]: { address: asAddress, privateKey: asPrivateKey },
+    };
+    const state = { ...defaults, initializingChannels: initializingChannelsA };
+    const action = actions.channel.ownCommitmentReceived(preFundCommitment2);
+    const updatedState = channelState.channelStateReducer(state, action);
+    const initializedChannel = states.getChannel(updatedState.state, channelId);
+
+    expect(initializedChannel).toBeUndefined(); // it doesn't transition
+    // TODO:
+    // this doesn't happen. The problem is that we don't hit `handleFirstCommit`
+    itSendsThisMessage(updatedState, SIGNATURE_FAILURE);
   });
 });
