+++ packages/wallet/src/redux/channel-state/reducer.ts
@@ -15,8 +15,15 @@ import {
   ReducerWithSideEffects,
   combineReducersWithSideEffects,
 } from '../../utils/reducer-utils';
-import { validCommitmentSignature } from '../../utils/signing-utils';
-import { showWallet, channelInitializationSuccess } from 'magmo-wallet-client/lib/wallet-events';
+import { validCommitmentSignature, signCommitment } from '../../utils/signing-utils';
+import {
+  showWallet,
+  channelInitializationSuccess,
+  signatureFailure,
+  signatureSuccess,
+  validationFailure,
+  validationSuccess,
+} from 'magmo-wallet-client/lib/wallet-events';
 import { CommitmentType } from 'fmg-core';
 import { StateWithSideEffects } from '../utils';
 import { ethers } from 'ethers';
@@ -33,35 +40,7 @@ export const channelStateReducer: ReducerWithSideEffects<states.ChannelState> = 
 
   const newState = { ...state };
   if (actions.isReceiveFirstCommitment(action)) {
-    // We manually select and move the initializing channel into the initializedChannelState
-    // before applying the combined reducer, so that the address and private key is in the
-    // right slot (by its channelId)
-    const channel = action.commitment.channel;
-    const channelId = channelID(channel);
-    if (newState.initializedChannels[channelId]) {
-      throw new Error('Channel already exists');
-    }
-    const initializingAddresses = new Set(Object.keys(newState.initializingChannels));
-    const ourAddress = channel.participants.find(addr => initializingAddresses.has(addr));
-    if (!ourAddress) {
-      return { state: newState };
-    }
-    const ourIndex = channel.participants.indexOf(ourAddress);
-
-    const { address, privateKey } = newState.initializingChannels[ourAddress];
-    delete newState.initializingChannels[ourAddress];
-    newState.initializedChannels[channelId] = states.waitForChannel({
-      address,
-      privateKey,
-      ourIndex,
-    });
-
-    // Since the wallet only manages one channel at a time, when it receives the first
-    // prefundsetup commitment for a channel, from the application, we set the
-    // activeAppChannelId accordingly.
-    // In the future, the application might need to specify the intended channel id
-    // for the action
-    newState.activeAppChannelId = channelId;
+    return handleFirstCommmit(state, action);
   }
 
   return combinedReducer(newState, action, {
@@ -84,10 +63,164 @@ const initializingChannels: ReducerWithSideEffects<states.InitializingChannelSta
       ...state,
       // We have to temporarily store the private key under the address, since
       // we can't know the channel id until both participants know their addresses.
-      [address]: states.waitForChannel({ address, privateKey }),
+      [address]: { address, privateKey },
     },
     sideEffects: { messageOutbox: channelInitializationSuccess(wallet.address) },
   };
+};
+
+type CommitmentReceived = actions.OwnCommitmentReceived | actions.OpponentCommitmentReceived;
+const handleFirstCommmit = (
+  state: states.ChannelState,
+  action: CommitmentReceived,
+): StateWithSideEffects<states.ChannelState> => {
+  // We manually select and move the initializing channel into the initializedChannelState
+  // before applying the combined reducer, so that the address and private key is in the
+  // right slot (by its channelId)
+  const channel = action.commitment.channel;
+  const channelId = channelID(channel);
+  if (state.initializedChannels[channelId]) {
+    throw new Error('Channel already exists');
+  }
+  const initializingAddresses = new Set(Object.keys(state.initializingChannels));
+  const ourAddress = channel.participants.find(addr => initializingAddresses.has(addr));
+  if (!ourAddress) {
+    return { state };
+  }
+  const ourIndex = channel.participants.indexOf(ourAddress);
+
+  const { address, privateKey } = state.initializingChannels[ourAddress];
+  delete state.initializingChannels[ourAddress];
+
+  // Since the wallet only manages one channel at a time, when it receives the first
+  // prefundsetup commitment for a channel, from the application, we set the
+  // activeAppChannelId accordingly.
+  // In the future, the application might need to specify the intended channel id
+  // for the action
+  state.activeAppChannelId = channelId;
+
+  const channelState = {
+    address,
+    privateKey,
+    ourIndex,
+  };
+
+  switch (action.type) {
+    case actions.OWN_COMMITMENT_RECEIVED:
+      const ownCommitment = action.commitment;
+
+      // check it's a PreFundSetupA
+      if (ownCommitment.commitmentType !== CommitmentType.PreFundSetup) {
+        // Since these checks are happening during a signature request we'll return a sig failure
+        return {
+          state,
+          sideEffects: {
+            messageOutbox: signatureFailure('Other', 'Expected a pre fund setup position.'),
+          },
+        };
+      }
+      if (ownCommitment.commitmentCount !== 0) {
+        return {
+          state,
+          sideEffects: { messageOutbox: signatureFailure('Other', 'Expected state count to 0.') },
+        };
+      }
+
+      if (ourAddress !== channelState.address) {
+        return {
+          state,
+          sideEffects: {
+            messageOutbox: signatureFailure(
+              'Other',
+              'Address provided does not match the one stored in the wallet.',
+            ),
+          },
+        };
+      }
+
+      const signature = signCommitment(ownCommitment, channelState.privateKey);
+      // if so, unpack its contents into the state
+      return {
+        state: states.setChannel(
+          state,
+          states.waitForPreFundSetup({
+            ...channelState,
+            libraryAddress: ownCommitment.channel.channelType,
+            channelId: channelID(ownCommitment.channel),
+            ourIndex: ownCommitment.channel.participants.indexOf(channelState.address),
+            participants: ownCommitment.channel.participants as [string, string],
+            channelNonce: ownCommitment.channel.nonce,
+            turnNum: 0,
+            lastCommitment: { commitment: ownCommitment, signature },
+            funded: false,
+          }),
+        ),
+        sideEffects: { messageOutbox: signatureSuccess(signature) },
+      };
+
+    case actions.OPPONENT_COMMITMENT_RECEIVED:
+      const opponentCommitment = action.commitment;
+
+      // all these checks will fail silently for the time being
+      // check it's a PreFundSetupA
+      if (opponentCommitment.commitmentType !== CommitmentType.PreFundSetup) {
+        return {
+          state,
+          sideEffects: {
+            messageOutbox: validationFailure('Other', 'Expected a prefund setup position'),
+          },
+        };
+      }
+      if (opponentCommitment.commitmentCount !== 0) {
+        return {
+          state,
+          sideEffects: {
+            messageOutbox: validationFailure('Other', 'Expected state count to be 0'),
+          },
+        };
+      }
+
+      const ourAddress2 = opponentCommitment.channel.participants[1];
+      const opponentAddress2 = opponentCommitment.channel.participants[0] as string;
+
+      if (!validCommitmentSignature(action.commitment, action.signature, opponentAddress2)) {
+        return {
+          state,
+          sideEffects: { messageOutbox: validationFailure('InvalidSignature') },
+        };
+      }
+
+      if (ourAddress2 !== channelState.address) {
+        return {
+          state,
+          sideEffects: {
+            messageOutbox: validationFailure(
+              'Other',
+              'Address provided does not match the one stored in the wallet.',
+            ),
+          },
+        };
+      }
+
+      // if so, unpack its contents into the state
+      return {
+        state: states.setChannel(
+          state,
+          states.waitForPreFundSetup({
+            ...channelState,
+            libraryAddress: opponentCommitment.channel.channelType,
+            channelId: channelID(opponentCommitment.channel),
+            ourIndex: opponentCommitment.channel.participants.indexOf(channelState.address),
+            participants: opponentCommitment.channel.participants as [string, string],
+            channelNonce: opponentCommitment.channel.nonce,
+            turnNum: 0,
+            lastCommitment: { commitment: action.commitment, signature: action.signature },
+            funded: false,
+          }),
+        ),
+        sideEffects: { messageOutbox: validationSuccess() },
+      };
+  }
 };
 
 const initializedChannels: ReducerWithSideEffects<states.InitializedChannelState> = (
@@ -109,8 +242,17 @@ const initializedChannels: ReducerWithSideEffects<states.InitializedChannelState
 
   const existingChannel = state[channelId];
   if (!existingChannel) {
-    // TODO:  This channel should really exist -- should we throw?
-    return { state };
+    // return a signature failure
+    return {
+      state,
+      sideEffects: {
+        messageOutbox: signatureFailure(
+          'Other',
+          `Wallet doesn't have an initialized channel for the provided commitment.
+          Expected a pre fund setup commitment.`,
+        ),
+      },
+    };
   }
 
   const { state: newState, sideEffects: outboxState } = initializedChannelStatusReducer(
